+++ CHANGELOG.md
@@ -12,6 +12,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Honeypot Captcha to detect and block bots without disrupting user experience.
 - Support of PHP 8.4
 
+### Fixed
+- Password validators are no longer used when password policy is disabled
+
 ## [1.0.0] - 2024-11-27
 This is the stable release of v1.0.0. No changes have been made since v1.0.0-rc.1.